requirements.txt
screeninfo
pillow
pynput